feat: Create master list fragment and display it when recipe card clicked.
The master list fragment will display the recipe selection screen.
A fragment is required because in tablet layout both this fragment
and a selected recipe step fragment will be displayed in a master-detail
flow.
The detail activity which is launched when recipe card is clicked will either
only display the master list fragment or display both the fragments.