fix: Ingredients list displays correctly. Only one ingredients text and space above ingredients is removed.
The error was because the RecipeStepsListFragment was being created twice,
first in the layout of DetailActivity and then dynamically through a fragment
transaction.
Created a container to hold the RecipeStepsListFragment and removed the
static fragment already in the layout to fix the error.
The recipe steps code was uncommented so the steps can also be seen.
Runs on API 28 emulator.