refactor: Pass in List of Steps and selected step index instead of just the selected Step.
We want to be able to change to a different Step when the next or
previous button is added later. To do that we may need to know what
all the steps are and what index is currently selected.
Shows correct step details on API 17 tablet.
Doesn't display any step details on API 28 phone emulator.