feat: Highlight selected item to indicate to user which recipe step he/she is on.
Change the background color of the selected step CardView to indicate which step
user is on. In two-pane layout this is useful because the user won't get
confused about which step details he/she is looking at.
However, the color is also changed on the single-pane layout.

This is done in onBindViewHolder(). Forcing onBindViewHolder() when
recipe step is clicked.