fix: Don't highlight first recipe step in single-pane layout.
In single-pane layout, since the user can't see recipe step
details until clicking on a step, we don't need to pre-select
the first recipe step to help user visually see which step they're at.

An error occurred in Android Studio where public methods in
RecipeStepsListFragment couldn't be detected by DetailActivity. This was
because the fragment was being declared as a Fragment, and Fragment does
not have the subclass RecipeStepsListFragment's methods.