fix: Set highlighted recipe step to the last step the user selected before an orientation change.
The user is at the same step he/she was before rotation, so the step
highlighted should match that step.
The RecipeStepsListFragment saved the highlighted position before rotation.
Then the DetailActivity stored an instance of RecipeStepsListFragment so
the Fragment could be saved before rotation, and therefore the state
contained inside that Fragment could be saved as well.

OutOfMemory error occurs with ExoPlayer after multiple rotations, crashes
API 17 tablet. Crashes on API 28 phone emulator.