fix: Retain recipe step details Fragment to prevent ExoPlayer OutOfMemory error after several orientation changes.
The problem was the following:
When the screen is rotated, the RecipeStepDetailsFragment is created twice.
Both of these created Fragments intialize ExoPlayer.
The first Fragment, however, calls its onDestroy and doesn't call onPause
or onStop. onPause is used to release the ExoPlayer. Since the first ExoPlayer
is never released, there is an OutOfMemory error after several screen rotations.
Reusing the RecipeStepDetailsActivity causes only one new Fragment to be created,
and that Fragment correctly releases ExoPlayer in onPause when there's a rotation.