refactor: Don't save list of recipe steps on orientation change.
Saving recipe steps when screen is rotated is unecessary because
the list will always be the same. Also, the Android documentation
says not to save large amounts of data with onSaveInstanceState.
Removed some logs and comments as well.

Also the previous commit 293441aca68e3d223c1b6b4024429e20c3b80c51 is
the same as the one before, was an accidental commit.