fix: Don't reset recipe ingredients text during periodic update.
We don't want to update the text during periodic widget updates, which in
this case is set to 30 minutes. Instead, just set a pending intent and
click listener to open MainActivity.
We still need to open MainActivity when the widget is first created,
which is why we do set the pending intent and click listener in onUpdate.